ASA tests passing tox
