--- test/unit/test_devices/device_mocks/asa/send_command/test_asa_device.py
@@ -1,357 +1,0 @@
-import pytest
-import os
-
-from unittest import mock
-
-from .device_mocks.asa import send_command, send_command_expect
-
-from pyntc.devices.asa_device import ASADevice, FileTransferError
-from pyntc.errors import CommandError, CommandListError
-
-
-BOOT_IMAGE = "c3560-advipservicesk9-mz.122-44.SE"
-
-
-class TestASADevice:
-    @mock.patch("pyntc.devices.asa_device.ConnectHandler")
-    def setup(self, api):
-
-        if not getattr(self, "device", None):
-            self.device = ASADevice("host", "user", "password")
-
-        # need to think if there should be an if before this...
-        self.device.native = api
-
-        # counts how many times we setup and tear down
-        if not getattr(self, "count_setup", None):
-            setattr(self, "count_setup", 0)
-
-        if not getattr(self, "count_teardown", None):
-            setattr(self, "count_teardown", 0)
-
-        self.device = ASADevice("host", "user", "password")
-        api.send_command_timing.side_effect = send_command
-        api.send_command_expect.side_effect = send_command_expect
-        self.device.native = api
-        self.count_setup += 1
-
-    def teardown(self):
-        # Reset the mock so that we do not have transient effects
-        self.device.native.reset_mock()
-
-        # count teardowns
-        self.count_teardown += 1
-
-    def test_config(self):
-        command = "hostname DATA-CENTER-FW"
-
-        result = self.device.config(command)
-        assert result is None
-
-        self.device.native.send_command_timing.assert_called_with(command)
-
-    def test_bad_config(self):
-        command = "asdf poknw"
-
-        with pytest.raises(CommandError, match=command):
-            self.device.config(command)
-
-    def test_config_list(self):
-        commands = ["crypto key generate rsa modulus 2048", "aaa authentication ssh console LOCAL"]
-        self.device.config_list(commands)
-
-        for cmd in commands:
-            self.device.native.send_command_timing.assert_any_call(cmd)
-
-    def test_bad_config_list(self):
-        commands = ["crypto key generate rsa modulus 2048", "lalala"]
-        results = ["ok", "Error: lalala"]
-
-        self.device.native.send_command_timing.side_effect = results
-
-        with pytest.raises(CommandListError, match=commands[1]):
-            self.device.config_list(commands)
-
-    def test_show(self):
-        command = "show running config"
-        result = self.device.show(command)
-
-        assert isinstance(result, str)
-        assert "interface" in result
-        assert "inspect" in result
-
-    def test_bad_show(self):
-        command = "show linux"
-        self.device.native.send_command_timing.return_value = "Error: linux"
-        with pytest.raises(CommandError):
-            self.device.show(command)
-
-    def test_show_list(self):
-        commands = ["show running config", "show startup-config"]
-
-        result = self.device.show_list(commands)
-        assert isinstance(result, list)
-        assert "console" in result[0]
-        assert "security-level" in result[1]
-
-        calls = list(mock.call(x) for x in commands)
-        self.device.native.send_command_timing.assert_has_calls(calls)
-
-    def test_bad_show_list(self):
-        commands = ["show badcommand", "show clock"]
-        results = ["Error: badcommand", "14:31:57.089 PST Tue Feb 10 2008"]
-
-        self.device.native.send_command_timing.side_effect = results
-
-        with pytest.raises(CommandListError, match="show badcommand"):
-            self.device.show_list(commands)
-
-    def test_save(self):
-        result = self.device.save()
-
-        assert result
-        # TODO: found that "copy run start" is the right command.
-        # https://www.networkstraining.com/cisco-asa-firewall-commands-cheat-sheet/
-        # TODO: ask Jacob
-        self.device.native.send_command_timing.assert_any_call("copy running-config startup-config")
-
-    @mock.patch("pyntc.devices.asa_device.FileTransfer", autospec=True)
-    def test_file_copy_remote_exists(self, mock_ft):
-        self.device.native.send_command_timing.side_effect = None
-        self.device.native.send_command_timing.return_value = "flash: /dev/null"
-
-        mock_ft_instance = mock_ft.return_value
-        mock_ft_instance.check_file_exists.return_value = True
-        mock_ft_instance.compare_md5.return_value = True
-
-        result = self.device.file_copy_remote_exists("source_file")
-
-        assert result
-
-    @mock.patch("pyntc.devices.asa_device.FileTransfer", autospec=True)
-    def test_file_copy_remote_exists_bad_md5(self, mock_ft):
-        self.device.native.send_command_timing.side_effect = None
-        self.device.native.send_command_timing.return_value = "flash: /dev/null"
-
-        mock_ft_instance = mock_ft.return_value
-        mock_ft_instance.check_file_exists.return_value = True
-        mock_ft_instance.compare_md5.return_value = False
-
-        result = self.device.file_copy_remote_exists("source_file")
-
-        assert not result
-
-    @mock.patch("pyntc.devices.asa_device.FileTransfer", autospec=True)
-    def test_file_copy_remote_exists_not(self, mock_ft):
-        self.device.native.send_command_timing.side_effect = None
-        self.device.native.send_command_timing.return_value = "flash: /dev/null"
-
-        mock_ft_instance = mock_ft.return_value
-        mock_ft_instance.check_file_exists.return_value = False
-        mock_ft_instance.compare_md5.return_value = True
-
-        result = self.device.file_copy_remote_exists("source_file")
-
-        assert not result
-
-    @mock.patch("pyntc.devices.asa_device.FileTransfer", autospec=True)
-    def test_file_copy(self, mock_ft):
-        self.device.native.send_command_timing.side_effect = None
-        self.device.native.send_command_timing.return_value = "flash: /dev/null"
-
-        mock_ft_instance = mock_ft.return_value
-        mock_ft_instance.check_file_exists.side_effect = [False, True]
-        self.device.file_copy("path/to/source_file")
-
-        mock_ft.assert_called_with(self.device.native, "path/to/source_file", "source_file", file_system="flash:")
-        mock_ft_instance.enable_scp.assert_any_call()
-        mock_ft_instance.establish_scp_conn.assert_any_call()
-        mock_ft_instance.transfer_file.assert_any_call()
-
-    @mock.patch("pyntc.devices.asa_device.FileTransfer", autospec=True)
-    def test_file_copy_different_dest(self, mock_ft):
-        self.device.native.send_command_timing.side_effect = None
-        self.device.native.send_command_timing.return_value = "flash: /dev/null"
-        mock_ft_instance = mock_ft.return_value
-
-        mock_ft_instance.check_file_exists.side_effect = [False, True]
-        self.device.file_copy("source_file", "dest_file")
-
-        mock_ft.assert_called_with(self.device.native, "source_file", "dest_file", file_system="flash:")
-        mock_ft_instance.enable_scp.assert_any_call()
-        mock_ft_instance.establish_scp_conn.assert_any_call()
-        mock_ft_instance.transfer_file.assert_any_call()
-
-    @mock.patch("pyntc.devices.asa_device.FileTransfer", autospec=True)
-    def test_file_copy_fail(self, mock_ft):
-        self.device.native.send_command_timing.side_effect = None
-        self.device.native.send_command_timing.return_value = "flash: /dev/null"
-        mock_ft_instance = mock_ft.return_value
-        mock_ft_instance.transfer_file.side_effect = Exception
-        mock_ft_instance.check_file_exists.return_value = False
-
-        with pytest.raises(FileTransferError):
-            self.device.file_copy("source_file")
-
-    def test_reboot(self):
-        self.device.reboot(confirm=True)
-        self.device.native.send_command_timing.assert_any_call("reload")
-
-    def test_reboot_with_timer(self):
-        self.device.reboot(confirm=True, timer=5)
-        self.device.native.send_command_timing.assert_any_call("reload in 5")
-
-    def test_reboot_no_confirm(self):
-        self.device.reboot()
-        assert not self.device.native.send_command_timing.called
-
-    # @mock.patch.object(ASADevice, "_get_file_system", return_value="bootflash:")
-    # def test_get_boot_options_show_bootvar(self, mock_boot):
-    #     self.device.native.send_command_timing.side_effect = None
-    #     self.device.native.send_command_timing.return_value = f"BOOT variable = bootflash:{BOOT_IMAGE}"
-    #     boot_options = self.device.get_boot_options()
-    #     assert boot_options == {"sys": BOOT_IMAGE}
-    #     self.device.native.send_command_timing.assert_called_with("show boot")
-
-    # @mock.patch.object(ASADevice, "_get_file_system", return_value="bootflash:")
-    # def test_get_boot_options_show_boot(self, mock_boot):
-    #     results = [
-    #         CommandError("show bootvar", "fail"),
-    #         f"BOOT path-list : bootflash:{BOOT_IMAGE}",
-    #     ]
-    #     self.device.native.send_command_timing.side_effect = results
-    #     boot_options = self.device.get_boot_options()
-    #     assert boot_options == {"sys": BOOT_IMAGE}
-    #     self.device.native.send_command_timing.assert_called_with("show boot")
-
-    # @mock.patch.object(ASADevice, "_get_file_system", return_value="bootflash:")
-    # def test_get_boot_options_show_run(self, mock_boot):
-    #     results = [
-    #         CommandError("show bootvar", "fail"),
-    #         CommandError("show bootvar", "fail"),
-    #         f"boot system flash bootflash:/{BOOT_IMAGE}",
-    #         "Directory of bootflash:/",
-    #     ]
-    #     self.device.native.send_command_timing.side_effect = results
-    #     boot_options = self.device.get_boot_options()
-    #     assert boot_options == {"sys": BOOT_IMAGE}
-    #     self.device.native.send_command_timing.assert_called_with("show run | inc boot")
-
-    # @mock.patch.object(ASADevice, "_get_file_system", return_value="flash:")
-    # @mock.patch.object(ASADevice, "get_boot_options", return_value={"sys": BOOT_IMAGE})
-    # @mock.patch.object(ASADevice, "config_list", return_value=None)
-    # def test_set_boot_options(self, mock_cl, mock_bo, mock_fs):
-    #     self.device.set_boot_options(BOOT_IMAGE)
-    #     mock_cl.assert_called_with(["no boot system", f"boot system flash:/{BOOT_IMAGE}"])
-
-    # @mock.patch.object(ASADevice, "_get_file_system", return_value="flash:")
-    # @mock.patch.object(ASADevice, "get_boot_options", return_value={"sys": BOOT_IMAGE})
-    # @mock.patch.object(ASADevice, "config_list", side_effect=[CommandError("boot system", "fail"), None])
-    # def test_set_boot_options_with_spaces(self, mock_cl, mock_bo, mock_fs):
-    #     self.device.set_boot_options(BOOT_IMAGE)
-    #     mock_cl.assert_called_with(["no boot system", f"boot system flash {BOOT_IMAGE}"])
-
-    # @mock.patch.object(ASADevice, "_get_file_system", return_value="flash:")
-    # def test_set_boot_options_no_file(self, mock_fs):
-    #     with pytest.raises(NTCFileNotFoundError):
-    #         self.device.set_boot_options("bad_image.bin")
-
-    # @mock.patch.object(ASADevice, "_get_file_system", return_value="flash:")
-    # @mock.patch.object(ASADevice, "get_boot_options", return_value={"sys": "bad_image.bin"})
-    # @mock.patch.object(ASADevice, "config_list", return_value=None)
-    # def test_set_boot_options_bad_boot(self, mock_cl, mock_bo, mock_fs):
-    #     with pytest.raises(CommandError):
-    #         self.device.set_boot_options(BOOT_IMAGE)
-    #         mock_bo.assert_called_once()
-
-    def test_backup_running_config(self):
-        filename = "local_running_config"
-        self.device.backup_running_config(filename)
-
-        with open(filename, "r") as f:
-            contents = f.read()
-
-        assert contents == self.device.running_config
-        os.remove(filename)
-
-    # rollback not implemented!
-    # def test_rollback(self):
-    #     self.device.rollback("good_checkpoint")
-    #     self.device.native.send_command_timing.assert_called_with("configure replace flash:good_checkpoint force")
-
-    # def test_bad_rollback(self):
-    #     # TODO: change to what the protocol would return
-    #     self.device.native.send_command_timing.return_value = "Error: rollback unsuccessful"
-    #     with pytest.raises(RollbackError):
-    #         self.device.rollback("bad_checkpoint")
-
-    def test_checkpoint(self):
-        self.device.checkpoint("good_checkpoint")
-        self.device.native.send_command_timing.assert_any_call("copy running-config good_checkpoint")
-
-    # not implemented
-    # def test_facts(self):
-    #     expected = {
-    #         "uptime": 413940,
-    #         "vendor": "cisco",
-    #         "uptime_string": "04:18:59:00",
-    #         "interfaces": ["FastEthernet0/0", "FastEthernet0/1"],
-    #         "hostname": "rtr2811",
-    #         "fqdn": "N/A",
-    #         "os_version": "15.1(3)T4",
-    #         "serial_number": "",
-    #         "model": "2811",
-    #         "vlans": [],
-    #         "cisco_ios_ssh": {"config_register": "0x2102"},
-    #     }
-    #     facts = self.device.facts
-    #     assert facts == expected
-
-    #     self.device.native.send_command_timing.reset_mock()
-    #     facts = self.device.facts
-    #     assert facts == expected
-
-    #     self.device.native.send_command_timing.assert_not_called()
-
-    def test_running_config(self):
-        expected = self.device.show("show running config")
-        assert self.device.running_config == expected
-
-    def test_starting_config(self):
-        expected = self.device.show("show startup-config")
-        assert self.device.startup_config == expected
-
-    def test_enable_from_disable(self):
-        self.device.native.check_enable_mode.return_value = False
-        self.device.native.check_config_mode.return_value = False
-        self.device.enable()
-        self.device.native.check_enable_mode.assert_called()
-        self.device.native.enable.assert_called()
-        self.device.native.check_config_mode.assert_called()
-        self.device.native.exit_config_mode.assert_not_called()
-
-    def test_enable_from_enable(self):
-        self.device.native.check_enable_mode.return_value = True
-        self.device.native.check_config_mode.return_value = False
-        self.device.enable()
-        self.device.native.check_enable_mode.assert_called()
-        self.device.native.enable.assert_not_called()
-        self.device.native.check_config_mode.assert_called()
-        self.device.native.exit_config_mode.assert_not_called()
-
-    def test_enable_from_config(self):
-        self.device.native.check_enable_mode.return_value = True
-        self.device.native.check_config_mode.return_value = True
-        self.device.enable()
-        self.device.native.check_enable_mode.assert_called()
-        self.device.native.enable.assert_not_called()
-        self.device.native.check_config_mode.assert_called()
-        self.device.native.exit_config_mode.assert_called()
-
-    def test_count_setup(self):
-        # This class is reinstantiated in every test, so the counter is reset
-        assert self.count_setup == 1
-
-    def test_count_teardown(self):
-        # This class is reinstantiated in every test, so the counter is reset
-        assert self.count_teardown == 0
